win32 installer: store artifact with tagged version
Added a simple script to call from AppVeyor to set the version correctly
on 'windows/snapcraft.iss'.

For this, 'determine_version' was extracted from 'setup.py'.

Appveyor has been reconfigured to store the non self signed
'snapcraft-installer.exe'.
