+++ setup.py
@@ -16,10 +16,11 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
-import subprocess
 import sys
 
 from setuptools import find_namespace_packages, setup
+
+from tools.version import determine_version
 
 
 def recursive_data_files(directory, install_directory):
@@ -28,39 +29,6 @@ def recursive_data_files(directory, install_directory):
         file_paths = [os.path.join(root, file_name) for file_name in file_names]
         data_files.append((os.path.join(install_directory, root), file_paths))
     return data_files
-
-
-def determine_version():
-    # Examples (git describe -> python package version):
-    # 4.1.1-0-gad012482d -> 4.1.1
-    # 4.1.1-16-g2d8943dbc -> 4.1.1.post16+g2d8943dbc
-    #
-    # For shallow clones or repositories missing tags:
-    # 0ae7c04
-
-    desc = (
-        subprocess.run(
-            ["git", "describe", "--always", "--long"],
-            check=True,
-            stdout=subprocess.PIPE,
-        )
-        .stdout.decode()
-        .strip()
-    )
-
-    split_desc = desc.split("-")
-    assert (
-        len(split_desc) == 3
-    ), f"Failed to parse Snapcraft git version description {desc!r}. Confirm that git repository is present and has the required tags/history."
-
-    version = split_desc[0]
-    distance = split_desc[1]
-    commit = split_desc[2]
-
-    if distance == "0":
-        return version
-
-    return f"{version}.post{distance}+git{commit[1:]}"
 
 
 # Common distribution data
